// frontend/src/app/models/interfaces/pack.interface.ts
export class PackInterface {
    // public productId: number;
    // public isbn: string;
    // public title: string;
    // public author: string;
    // public description: string;
    // public publisher: string;
    // public publishedYear: number;
    // public price: number;

    id?: number;
    name?: string;
    description?: string;
    price?: number;
    mass?: number;
    volume?: number;
    pickUpAddress?: string;
    pickUpCity?: string;
    pickUpZipCode?: string;
    pickUpCountry?: string;
    pickUpDate?: string;
}